+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 
 ### Added
 
+* Add bindings for the Web Locks API to `web-sys`.
+  [#3604](https://github.com/rustwasm/wasm-bindgen/pull/3604)
+
 * Add bindings for `ViewTransition` to `web-sys`.
   [#3598](https://github.com/rustwasm/wasm-bindgen/pull/3598)